fix: ListDetail Scaffold for CardDetailScreen
