+++ feature/card/src/main/java/com/jetapps/jettaskboard/carddetailscomponents/components/EditTextCard.kt
@@ -1,4 +1,4 @@
-package com.jetapps.jettaskboard.carddetailscomponents
+package com.jetapps.jettaskboard.carddetailscomponents.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Row
